requirements.txt
fastapi
uvicorn
python-multipart
requests
google-generativeai
pdf2image
Pillow
pydantic
python-dotenv